feat(hera): rebased on bin/ location change
