--- bin/kona-exex/src/kona.rs
@@ -1,59 +1,0 @@
-use std::sync::Arc;
-
-use eyre::{bail, Result};
-use reth_exex::{ExExContext, ExExEvent};
-use reth_node_api::FullNodeComponents;
-use superchain_registry::RollupConfig;
-use tracing::{debug, info};
-
-use crate::cli::KonaArgsExt;
-
-/// The Kona Execution Extension.
-#[derive(Debug)]
-#[allow(unused)]
-pub(crate) struct KonaExEx<Node: FullNodeComponents> {
-    /// The rollup configuration
-    cfg: Arc<RollupConfig>,
-    /// The context of the Execution Extension
-    ctx: ExExContext<Node>,
-}
-
-#[allow(unused)]
-impl<Node: FullNodeComponents> KonaExEx<Node> {
-    /// Creates a new instance of the Kona Execution Extension.
-    pub async fn new(ctx: ExExContext<Node>, args: KonaArgsExt, cfg: Arc<RollupConfig>) -> Self {
-        Self { ctx, cfg }
-    }
-
-    /// Wait for the L2 genesis L1 block (aka "origin block") to be available in the L1 chain.
-    async fn wait_for_l2_genesis_l1_block(&mut self) -> Result<()> {
-        loop {
-            if let Some(notification) = self.ctx.notifications.recv().await {
-                if let Some(committed_chain) = notification.committed_chain() {
-                    let tip = committed_chain.tip().block.header().number;
-                    // TODO: commit the chain to a local buffered provider
-                    // self.chain_provider.commit_chain(committed_chain);
-
-                    if let Err(err) = self.ctx.events.send(ExExEvent::FinishedHeight(tip)) {
-                        bail!("Critical: Failed to send ExEx event: {:?}", err);
-                    }
-
-                    if tip >= self.cfg.genesis.l1.number {
-                        break Ok(());
-                    } else {
-                        debug!(target: "kona", "Chain not yet synced to rollup genesis. L1 block number: {}", tip);
-                    }
-                }
-            }
-        }
-    }
-
-    /// Starts the Kona Execution Extension loop.
-    pub async fn start(mut self) -> Result<()> {
-        // Step 1: Wait for the L2 origin block to be available
-        self.wait_for_l2_genesis_l1_block().await?;
-        info!(target: "kona", "Chain synced to rollup genesis");
-
-        todo!("init pipeline and start processing events");
-    }
-}
